% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_presto.R
\name{connectPresto}
\alias{connectPresto}
\title{Connect to server using RPresto}
\usage{
connectPresto(server_type = "nm", db_type = "dbi",
  schema_name = NA_character_)
}
\arguments{
\item{server_type}{string}

\item{db_type}{string}

\item{schema_name}{string}
}
\description{
Connect to server using RPresto
}
