% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{fillNAs}
\alias{fillNAs}
\title{fillNAs: fill NAs with the most frequent value(for numeric) or default value(for category)}
\usage{
fillNAs(dt, col.name, default.cat = "None", default.num = NA)
}
\arguments{
\item{dt}{data.table}

\item{col.name}{}
}
\description{
fillNAs: fill NAs with the most frequent value(for numeric) or default value(for category)
}
\examples{


}
