% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\docType{data}
\name{pkgs_required}
\alias{pkgs_required}
\title{Load most frequently used packages}
\format{An object of class \code{character} of length 16.}
\usage{
pkgs_required
}
\arguments{
\item{pkgs}{packages to be loaded}

\item{install}{if the package is not installed, install it or not}
}
\value{
vector, if a package is sucessfully loaded
}
\description{
Load most frequently used packages
}
\examples{
load_packages()

}
\keyword{datasets}
