% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{showNAInfo}
\alias{showNAInfo}
\title{showNAInfo: show columns with NA value}
\usage{
showNAInfo(df, na.prop = 0)
}
\arguments{
\item{df}{data.frame}

\item{na.prop}{numeric, default = 0.0}
}
\value{
data.table, show columns with prop ('= #na_cnt/#total_cnt) equal or larger than na.prop
}
\description{
showNAInfo: show columns with NA value
}
\examples{
showNAInfo(flights)

}
