% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_RYLin.R
\name{updateRYLin}
\alias{updateRYLin}
\title{Update the library
(borrowed from laresbernardo/lares)
This function lets the user update from repository or local source.}
\usage{
updateRYLin(local = FALSE, force = FALSE, restart = FALSE)
}
\arguments{
\item{local}{Boolean. Install package with local files (TRUE) or Github repository}

\item{force}{Boolean. Force install if needed}

\item{restart}{Boolean. Restart session after re-installing the library}
}
\description{
Update the library
(borrowed from laresbernardo/lares)
This function lets the user update from repository or local source.
}
